mgr/rbd_support: generalize schedule reusable code
Also fix the bug in recursive list, when global schedule is not
set and listing a subtree without schedules.
